+++ lib/styles/gradients.dart
@@ -1,0 +1,13 @@
+import 'package:anchwatt/styles/colors.dart';
+import 'package:flutter/material.dart';
+
+// Level max
+const LinearGradient gradientLevelMax = LinearGradient(
+  begin: Alignment.topLeft,
+  end: Alignment.bottomRight,
+  colors: [
+    colorLevelMaxGoldHighlight,
+    colorLevelMaxGold,
+    colorLevelMaxGoldDeep,
+  ],
+);